Added more description for notice.
Some descriptions are not enough, and we added to the more description. This project is only meet these requirements and you can run it.  When you don't meet these requirements, then you have not to run this program and that is why you should meet these requirements to run it.,